Refactor Game scene file
