+++ src/generators/coins.js
