+++ src/generators/platforms.js
