+++ src/preloaders/spriteSheets.js
