+++ src/scenes/Welcome.js
@@ -1,7 +1,7 @@
 /* eslint-disable class-methods-use-this */
 import Phaser from 'phaser';
 import gameState from '../state/state';
-import images from '../helpers/images';
+import images from '../preloaders/images';
 
 class WelcomeScene extends Phaser.Scene {
   constructor() {
